+++ CHANGELOG.md
@@ -153,6 +153,8 @@ API-breaking change is listed.
   `difference_insert`, `difference_insert_subseteq`. (by Hai Dang)
 - Tweak reduction on `gset`, so that `cbn` matches the behavior by `simpl` and
   does not unfold `gset` operations. (by Paolo G. Giarrusso, BedRock Systems)
+- Add `get_head` tactic to determine the syntactic head of a function
+  application term.
 
 The following `sed` script should perform most of the renaming
 (on macOS, replace `sed` by `gsed`, installed via e.g. `brew install gnu-sed`).